adds rating to review

+++ app/src/main/java/edu/dixietech/lukassimonson/moovies/features/review/model/repository/ReviewRepository.kt
@@ -6,5 +6,5 @@ import edu.dixietech.lukassimonson.moovies.shared.domain.Review
 interface ReviewRepository {
     suspend fun getMovieReview(id: Int): Movie?
     suspend fun saveMovieReview(movie: Movie, review: Review)
-    suspend fun saveMovieRating(movie: Movie, rating: Double)
+    suspend fun saveMovieRating(movie: Movie, rating: Int)
 }
